+++ ServiceAbonents/Dtos/TarrifDto.cs
@@ -3,9 +3,9 @@
     public class TarrifDto
     {
         public string TariffName { get; set; }
-        public short RemainGb { get; set; }
-        public short RemainMin { get; set; }
-        public short RemainSMS { get; set; }
+        public decimal RemainGb { get; set; }
+        public decimal RemainMin { get; set; }
+        public decimal RemainSMS { get; set; }
         public bool UnlimVideo { get; set; }
         public bool UnlimSocials { get; set; }
         public bool UnlimMusic { get; set; }
